+++ Assets/Scripts/Common/TemplateAnimation.cs
@@ -1,11 +1,11 @@
 using System;
 using UnityEngine;
 
-namespace TextAnimation
+namespace TextAnimation.Common
 {
 
-    [CreateAssetMenu(fileName = nameof(WaveAnimation), menuName = "TextAnimation/Movement/" + nameof(WaveAnimation), order = 1)]
-    internal class WaveAnimation : BaseAnimationObject
+    [CreateAssetMenu(fileName = nameof(TemplateAnimation), menuName = "TextAnimation/Unfiltered/" + nameof(TemplateAnimation), order = 1)]
+    internal class TemplateAnimation : BaseAnimationObject
     {
         #region Internal
         /// <summary>
@@ -13,9 +13,9 @@ namespace TextAnimation
         /// </summary>
         /// <param name="mesh">The mesh that will be given to modify.</param>
         /// <returns>The modified mesh.</returns>
-        internal override Mesh DoEffect(Mesh mesh)
+        public override Mesh DoEffect(Mesh mesh)
         {
-            return WaveAnimation(mesh);
+            return mesh; // Currently, does
         }
 
         #endregion
