+++ Assets/Scripts/Common/TextAnimation.Common.asmdef
@@ -1,6 +1,6 @@
 {
-    "name": "TextAnimation",
-    "rootNamespace": "TextAnimation",
+    "name": "TextAnimation.Common",
+    "rootNamespace": "TextAnimation.Common",
     "references": [
         "GUID:6055be8ebefd69e48b49212b09b47b2f"
     ],
