+++ Assets/Scripts/Examples/StaticColourPurple.asset
