+++ Assets/Scripts/Examples/StaticColourRed.asset
